refactor: split tests for comparing helpers
